Planning_2.1_branch_Gemini/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
streamlit-aggrid>=0.3.4
PyGithub>=2.1.1
